address Windows VS10 warnings

+++ maptk/core/descriptor.h
@@ -67,9 +67,9 @@ public:
   /// Return the descriptor as a vector of doubles
   std::vector<double> as_double() const
   {
-    const unsigned length = this->size();
+    const size_t length = this->size();
     std::vector<double> double_data(length);
-    for(unsigned i=0; i<length; ++i)
+    for(size_t i=0; i<length; ++i)
     {
       double_data[i] = static_cast<double>(this->raw_data()[i]);
     }
